+++ releases.md
@@ -1,5 +1,9 @@
 # Releases
 
+## Unreleased
+
+  - Improve compatibility of rackup handler w.r.t. sinatra.
+
 ## v0.47.8
 
   - Fix Falcon Supervisor implementation: due to invalid code, it was unable to start.